requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
altair=3.2.0=py37_0
asgiref=3.2.7=py_0
astunparse=1.6.3=pypi_0
attrs=19.3.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.1.5=py_0
boto=2.49.0=py37_0
boto3=1.9.66=py37_0
botocore=1.12.189=py_0
brotlipy=0.7.0=py37h7b6447c_1000
ca-certificates=2020.1.1=0
cachetools=4.1.0=py_1
certifi=2020.6.20=py37_0
cffi=1.14.0=py37he30daa8_1
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cryptography=2.9.2=py37h1ba5d50_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py37_0
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
django=3.0.3=py_0
docopt=0.6.2=py37_0
docutils=0.16=py37_1
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
funcy=1.14=py_0
future=0.18.2=py37hc8dfbb8_1
gast=0.3.3=pypi_0
gensim=3.8.0=py37h962f231_0
glib=2.63.1=h3eb4bd4_1
gmp=6.1.2=h6c8ec71_1
google-api-core=1.17.0=py37_0
google-auth=1.14.1=py_0
google-auth-oauthlib=0.4.1=pypi_0
google-cloud-core=1.3.0=py_0
google-cloud-storage=1.28.0=py_0
google-pasta=0.2.0=pypi_0
google-resumable-media=0.5.0=py_1
googleapis-common-protos=1.51.0=py37_2
grpcio=1.29.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=pypi_0
icu=58.2=he6710b0_3
idna=2.9=py_1
importlib-metadata=1.6.1=py37_0
importlib_metadata=1.6.1=0
intel-openmp=2020.1=217
ipykernel=5.3.0=py37h5ca1d4c_0
ipython=7.15.0=py37_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.17.1=py37_0
jinja2=2.11.2=py_0
jmespath=0.9.4=py_0
joblib=0.15.1=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py37_0
jupyter_client=6.1.3=py_0
jupyter_core=4.6.3=py37_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=py37hfd86e86_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h7b6447c_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.12.3=hd408876_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_0
markdown=3.2.2=pypi_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py37hef1b27d_0
mistune=0.8.4=py37h7b6447c_0
mkl=2020.1=217
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.1.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
nbconvert=5.6.1=py37_0
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
nltk=3.5=py_0
notebook=6.0.3=py37_0
numexpr=2.7.1=py37h0da4684_1
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1g=h7b6447c_0
opt-einsum=3.2.1=pypi_0
packaging=20.4=py_0
pandas=1.0.5=py37h0573a6f_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.1.2=py37hb39fc2d_0
pip=20.1.1=py37_1
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.5=py_0
protobuf=3.12.3=py37he6710b0_0
ptyprocess=0.6.0=py37_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pygments=2.6.1=py_0
pyldavis=2.1.2=py_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.16.0=py37h7b6447c_0
pysocks=1.7.1=py37_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.5.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pytz=2020.1=py_0
pyzmq=18.1.1=py37he6710b0_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2020.5.14=py37h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=py_0
s3transfer=0.1.13=py37_0
scikit-learn=0.23.1=pypi_0
scipy=1.4.1=py37h0b6359f_0
send2trash=1.5.0=py37_0
setuptools=47.3.1=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sklearn=0.0=pypi_0
smart_open=2.0.0=py_0
sqlite=3.32.3=h62c20be_0
sqlparse=0.3.1=py_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.6.0.post3=pypi_0
tensorflow=2.2.0=pypi_0
tensorflow-estimator=2.2.0=pypi_0
tensorflow-hub=0.8.0=pypi_0
tensorflow-text=2.2.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.6.1=py37_cu102
tornado=6.0.4=py37h7b6447c_1
tqdm=4.46.1=py_0
traitlets=4.3.3=py37_0
urllib3=1.25.9=py_0
wcwidth=0.2.4=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.1=he6710b0_3
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
